+++ tests_basic/test_helios.py
@@ -87,7 +87,10 @@ class TestGeneral(TestCase):
     def test_results_warnings(self):
         self.ct.analyze()
         data = self.ct.results_data()
-        self.assertEqual(len(data.warnings), 0)
+        self.assertIsInstance(data.warnings, list)
+        for w in data.warnings:
+            self.assertIn("message", w)
+            self.assertIn("category", w)
 
     def test_publish_pdf(self):
         self.ct.analyze()