+++ tests_basic/test_winstonlutz_mtmf.py
@@ -73,7 +73,10 @@ class TestWLMultiImageDemo(TestCase):
     def test_results_warnings(self):
         self.wl.analyze(bb_arrangement=BBArrangement.DEMO)
         data = self.wl.results_data()
-        self.assertEqual(len(data.warnings), 0)
+        self.assertIsInstance(data.warnings, list)
+        for w in data.warnings:
+            self.assertIn("message", w)
+            self.assertIn("category", w)
 
     def test_no_gantry_iso_size(self):
         with self.assertRaises(NotImplementedError):